Adding garo and garo master, finishing intrinsic change, minor fixes
Adding new intrinsic resistances for monsters involved expanding the size of the various variables that hold each monster's resistance flags.

Adding drain and sickness resistance meant reordering the values found in prop.h

Engrave.c: A minor fix to Buer's title, unrelated to garo and intriniscs

monst.c: added drain and sickness resitance tags to creatures that should have them. Added garo and garo master. Moved the Binder entry so that its position in this list matches its position in the tiles file and on the start screen.

rumors.c: outgmaster is a chopped-down version of the oracles script, such that it provides one unheard consultation.

monsters.txt: added tiles for garo and garo master, added non-placeholder tile for Exile(Binder)

+++ dnethack-3.4.3/include/monst.h
@@ -60,7 +60,7 @@ struct monst {
 #define M_AP_MONSTER	3	/* a monster */
 
 	schar mtame;		/* level of tameness, implies peaceful */
-	unsigned short mintrinsics;	/* low 8 correspond to mresists */
+	unsigned int mintrinsics;	/* low 10 correspond to mresists */
 	int mspec_used;		/* monster's special ability attack timeout */
 
 	Bitfield(female,1);	/* is female */
