+++ dnethack-3.4.3/include/permonst.h
@@ -53,8 +53,8 @@ struct permonst {
 	short		pxlth;			/* length of extension */
 	uchar		msound;			/* noise it makes (6 bits) */
 	uchar		msize;			/* physical size (3 bits) */
-	uchar		mresists;		/* resistances */
-	uchar		mconveys;		/* conveyed by eating */
+	unsigned int	mresists;		/* resistances */
+	unsigned int	mconveys;		/* conveyed by eating */
 	unsigned long	mflags1,		/* boolean bitflags */
 			mflags2;		/* more boolean bitflags */
 	unsigned long	mflags3;		/* yet more boolean bitflags *///increased from short to long
